Realizada a parametrização para excluir a avaliação inicial de um fornecedor que ainda não tenha sido avaliado anteriormente

+++ Dominio/Servico/AvaliaCriterioAvaliacaoServico.cs
@@ -29,7 +29,7 @@ namespace Dominio.Servico
 
                     var historico = new HistoricoCriterioAvaliacao
                     {
-                        IdCriterioAvaliacao = avaliaCriterioAvaliacao.IdCriterioAvaliacao,
+                        IdCriterioAvaliacao = (int)avaliaCriterioAvaliacao.IdCriterioAvaliacao,
                         Nota = avaliaCriterioAvaliacao.NotaAvaliacao.Value
                     };
